tweak: add general solution for propmpt modal instead of custom one.
