+++ app/vars/ipcConsts.ts
@@ -4,6 +4,7 @@ enum IpcChannel {
   TOGGLE_AUTO_START_ON_SYSTEM_LAUNCH = 'TOGGLE_AUTO_START_ON_SYSTEM_LAUNCH',
   IS_AUTO_START_ON_SYSTEM_LAUNCH_ENABLED_REQUEST = 'IS_AUTO_START_ON_SYSTEM_LAUNCH_ENABLED_REQUEST',
   SET_NODE_PORT = 'SET_NODE_PORT',
+  PROMPT_MODAL_REQUEST = 'PROMPT_MODAL_REQUEST',
   // Theme
   GET_OS_THEME_COLOR = 'GET_OS_THEME_COLOR',
   OPEN_BROWSER_VIEW = 'OPEN_BROWSER_VIEW',
@@ -59,7 +60,6 @@ enum IpcChannel {
   W_M_SPAWN_TX = 'W_M_SPAWN_TX',
   W_M_SEND_TX = 'W_M_SEND_TX',
   W_M_UPDATE_TX_NOTE = 'W_M_UPDATE_TX_NOTE',
-  W_M_APPROVE_ADD_WALLET_REQUEST = 'W_M_APPROVE_ADD_WALLET_REQUEST',
   W_M_APPROVE_ADD_WALLET_RESPONSE = 'W_M_APPROVE_ADD_WALLET_RESPONSE',
   // Transactions Manager
   T_M_UPDATE_ACCOUNT = 'T_M_UPDATE_ACCOUNT',
