+++ shared/ipcMessages.ts
@@ -108,12 +108,6 @@ export type AppLogs = {
 
 export type ShowFileRequest = { filePath?: string; isLogFile?: boolean };
 
-export type ImportWalletWarningRequest = {
-  isDuplicateName: boolean;
-  isDuplicateWallet: boolean;
-  isDuplicatePath: boolean;
-};
-
 export enum AddWalletResponseType {
   Error,
   DuplicateNotAllowed,
